requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv>=1.0.0
motor==3.3.2
pymongo==4.6.1
httpx==0.26.0
google-genai>=1.2.0
boto3>=1.34.0
